QUESTmetrics/QUESTmetrics/backend/requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-cors
flask-jwt-extended
python-dotenv
marshmallow-sqlalchemy
datetime
names
sklearn